requirements.txt
uvicorn
fastapi
python-multipart
psycopg2-binary
sqlalchemy
PyYAML
python-jose